useBlockActions hook (#2172)
